+++ src/components/CheckBox/CheckBox.test.tsx
@@ -138,4 +138,9 @@ describe('CheckBox', () => {
       )
     ).toEqual(true);
   });
+
+  test('CheckBox with toggle is not readOnly', () => {
+    const wrapper = mount(<CheckBox toggle label="test label" />);
+    expect(wrapper.find('input').prop('readOnly')).toBeFalsy();
+  });
 });